Move @Model ExerciseEntity and its extensions to production. Also changed the platform to .macOS(.v14) as it's the minimum requirement to run SwiftData
