+++ WorkoutTrack/Package.swift
@@ -7,7 +7,7 @@ let package = Package(
     name: "WorkoutTrack",
     platforms: [
         .iOS(.v17),
-        .macOS(.v13)
+        .macOS(.v14)
     ],
     products: [
         // Products define the executables and libraries a package produces, making them visible to other packages.
